+++ sdk/AttachmentPushService.m
@@ -265,6 +265,12 @@ NSString * const kAttachmentBackgroundSessionIdentifier = @"mil.nga.mage.backgro
     attachment.remoteId = [response valueForKey:@"id"];
     attachment.name = [response valueForKey:@"name"];
     attachment.url = [response valueForKey:@"url"];
+
+    // processing fields - reads from servers json response but sets the attributes in xcode
+    attachment.processingStatus = [response valueForKey:@"processingStatus"];
+    attachment.processingMessage = [response valueForKey:@"processingMessage"];
+    attachment.processingHook = [response valueForKey:@"processingHook"];
+
     attachment.taskIdentifier = nil;
     NSString *dateString = [response valueForKey:@"lastModified"];
     if (dateString != nil) {
@@ -292,6 +298,12 @@ NSString * const kAttachmentBackgroundSessionIdentifier = @"mil.nga.mage.backgro
             }
             
             [NSNotificationCenter.defaultCenter postNotificationName:@"AttachmentPushed" object:nil];
+        }];
+    } else if (attachment.processingStatus != nil) {
+        // Server is tracking status
+        __weak __typeof__(self) weakSelf = self;
+        [context MR_saveToPersistentStoreWithCompletion:^(BOOL contextDidSave, NSError * _Nullable error) {
+            [weakSelf.pushTasks removeObject:[NSNumber numberWithLong:task.taskIdentifier]];
         }];
     } else {
         // try again